Move components to Project page folder
- Create modular media sass file for Projects page
- Move InputField and TodoList and TodoItem components to Project page folder
- Fix slightly project page styling
