+++ src/pages/Project/Project.module.scss
@@ -53,7 +53,8 @@
 
 .project_container {
     //width: 100vw;
-    height: 100vh;
+    //height: 100vh;
+    height: 60%;
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -68,7 +69,7 @@
     display: flex;
     justify-content: space-between;
     width: 100%;
-    height: 60%;
+    height: 100%;
     padding: 0 5%;
     gap: 2rem;
 }
@@ -174,5 +175,91 @@
 
     &.dragging_over {
         background-color: rgb(174, 75, 5);
+    }
+}
+
+
+// TodoList content
+.todolist_container {
+    list-style: none;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    align-self: flex-start;
+    gap: .75rem;
+    margin: 0;
+    padding: 0;
+    //padding: .25rem;
+    //overflow-y: auto;
+
+    .todo_item {
+        width: 90%;
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        background-color: orange;
+        border-radius: .5rem;
+        padding: .25rem 1rem;
+        cursor: grab;
+        transition: all .2s ease-in-out;
+
+        &.todo_item_priority {
+            background-color: rgb(255, 85, 0);
+            box-shadow: 0 0 1rem .25rem rgb(255, 85, 0);
+            scale: 1.01;
+
+            &:hover {
+                box-shadow: 0 0 1.25rem rgb(124, 29, 0);
+            }
+        }
+
+        &:hover {
+            scale: 1.05;
+            box-shadow: 0 0 .35rem black;
+        }
+
+        &.dragging {
+            box-shadow: 0 0 2rem .35rem rgb(44, 44, 44);
+        }
+
+        &.drop_animating {
+            transition: all 1s ease-in-out;
+            opacity: .5;
+        }
+
+        .todo_item_edit_form {
+            width: 75%;
+            display: flex;
+            justify-content: flex-start;
+            gap: .25rem;
+
+            .todo_item_edit_input {
+                width: 75%;
+            }
+
+            button {
+                width: 15%;
+            }
+        }
+
+        .todo_item_text {
+            color: green;
+
+            // &.todo_item_text_priority {
+            //     //text-decoration: line-through;
+            //     background-color: grey;
+            // }
+        }
+
+        .todo_item_btn_container {
+            display: flex;
+            gap: .75rem;
+
+            div {
+                cursor: pointer;
+            }
+        }
     }
 }
