+++ src/pages/Project/TodoItem.tsx
@@ -2,9 +2,8 @@ import React, { useContext, useEffect, useRef, useState } from "react";
 import { Draggable, DraggableStateSnapshot } from "react-beautiful-dnd";
 
 import { Todo } from "../../models/model";
-//import { TodoListsDispatchContext } from "../../App";
 import { TodoListsDispatchContext } from '../../pages/Project/Project';
-import styles from "./TodoList.module.scss";
+import styles from "./Project.module.scss";
 
 type Props = {
     todo: Todo;
