+++ src/pages/Project/_media.scss
@@ -1,0 +1,6 @@
+// Media query for mobile content (portrait)
+@media (max-width: 500px) {
+    .input {
+        width: 80%;
+    }
+}
